Add ability to compile and run react controls in a webViewPanel (#7)
Implement basic webview support. This should allow us to iterate on the datascience controls using react.
This commit creates a history pane that just looks like the default "react" page but works inside of a webview panel.
Additionally it still makes it possible to load the index.html into a browser for debugging the react code standalone.